+++ src/main/java/net/silentchaos512/gear/util/DataResource.java
@@ -1,0 +1,13 @@
+package net.silentchaos512.gear.util;
+
+import net.minecraft.resources.ResourceLocation;
+
+import java.util.function.Function;
+
+// Deprecated: use api package class instead
+@Deprecated
+public class DataResource<T> extends net.silentchaos512.gear.api.util.DataResource<T> {
+    public DataResource(ResourceLocation id, Function<ResourceLocation, T> getter) {
+        super(id, getter);
+    }
+}